# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Senne Platform Documentation'
  version: 1.0.0
paths:
...
  '/api/treatment/exams/{uuid}/{numatendimento}':
    get:
      tags:
        - 'List Attedance'
      summary: Summary
      operationId: 4ac4a662881ea2b6afe6713e37b56ca3
      parameters:
        -
          name: uuid
          in: query
          required: true
          schema:
            type: string
        -
          name: numatendimento
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
  '/api/treatment/exams/{uuid}/{startdate}/{finaldate}':
    get:
      tags:
        - 'List Attedance Hospital - Date'
      summary: Summary
      operationId: 738d67eb937551e6abbb8bf3a59e37e6
      parameters:
        -
          name: uuid
          in: query
          required: true
          schema:
            type: string
        -
          name: startdate
          in: query
          required: true
          schema:
            type: string
        -
          name: finaldate
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
  '/api/treatment/details/{uuid}/{atendimento}':
    get:
      tags:
        - 'List Attedance Details'
      summary: Summary
      operationId: 2dca0a5f5c4a423abf302d4419afd966
      parameters:
        -
          name: uuid
          in: query
          required: true
          schema:
            type: string
        -
          name: atendimento
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
  '/api/treatment/report/{uuid}/{r_id}':
    get:
      tags:
        - 'List Principal Report'
      summary: Summary
      operationId: f21d8f8f518a4048623fad6a5830de77
      parameters:
        -
          name: uuid
          in: query
          required: true
          schema:
            type: string
        -
          name: r_id
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
  /api/list/procedencia:
    get:
      tags:
        - 'List All Hospitals '
      summary: Summary
      operationId: 85c166eb0e5d619b933b7b6e88cfba26
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
...